+++ src/Blocks/components/form/assets/form.js
@@ -433,6 +433,11 @@ export class Form {
 				// Set global msg.
 				this.utils.setGlobalMsg(formId, message, status, data);
 
+				// This will be changed in the next release.
+				if (Boolean(this.state.getStateFormElement(formId)?.getAttribute(this.state.getStateAttribute('formHideOnSuccess')))) {
+					this.state.getStateFormElement(formId).classList.add(this.state.getStateSelector('isHidden'));
+				}
+
 				// Remove local storage for prefill.
 				if (this.state.getStateEnrichmentIsUsed()) {
 					this.enrichment.deleteLocalStorage(this.state.getStateEnrichmentFormPrefillStorageName(formId));