+++ src/Rest/Routes/Integrations/Mailer/FormSubmitMailerRoute.php
@@ -106,6 +106,10 @@ class FormSubmitMailerRoute extends AbstractFormSubmit
 			);
 		}
 
+		$this->getFormSubmitMailer()->sendFallbackIntegrationEmail(
+			$formDetails
+		);
+
 		return \rest_ensure_response(
 			UtilsApiHelper::getApiErrorPublicOutput(
 				$this->labels->getLabel($label, $formId),