+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [4.0.6]
+
+### Updated
+- `Eightshift-forms-utils` to the latest version `2.0.2`.
+- result output item now supports conditional logic with both single and duplicate values.
+
+### Added
+- email fallbacks now supports request IP sent from the server, the data is anonymized.
+- new form setting that will hide form once it is submitted with success.
+- new `esFormsROISF` shortcode that can be used inside the result output item to show the form again.
+
 ## [4.0.5]
 
 ### Updated
@@ -450,6 +461,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
+[4.0.6]: https://github.com/infinum/eightshift-forms/compare/4.0.5...4.0.6
 [4.0.5]: https://github.com/infinum/eightshift-forms/compare/4.0.4...4.0.5
 [4.0.4]: https://github.com/infinum/eightshift-forms/compare/4.0.3...4.0.4
 [4.0.3]: https://github.com/infinum/eightshift-forms/compare/4.0.2...4.0.3